# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AdvisorySource": {
        "additionalProperties": false,
        "description": "A recently observed source used only for non-authoritative live context.",
        "properties": {
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "verified_at": {
            "format": "date-time",
            "title": "Verified At",
            "type": "string"
          }
        },
        "required": [
          "title",
          "url",
          "verified_at"
        ],
        "title": "AdvisorySource",
        "type": "object"
      },
...
      "CatQuestionnaire": {
        "additionalProperties": false,
        "properties": {
          "activity_level": {
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Activity Level",
            "type": "string"
          },
          "affection_preference": {
            "enum": [
              "INDEPENDENT",
              "BALANCED",
              "AFFECTIONATE"
            ],
            "title": "Affection Preference",
            "type": "string"
          },
          "children_present": {
            "default": false,
            "title": "Children Present",
            "type": "boolean"
          },
          "existing_pets": {
            "items": {
              "enum": [
                "DOG",
                "CAT"
              ],
              "type": "string"
            },
            "title": "Existing Pets",
            "type": "array",
            "uniqueItems": true
          },
          "experience": {
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "EXPERT"
            ],
            "title": "Experience",
            "type": "string"
          },
          "grooming_tolerance": {
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Grooming Tolerance",
            "type": "string"
          },
          "home_size": {
            "enum": [
              "SMALL",
              "MEDIUM",
              "LARGE"
            ],
            "title": "Home Size",
            "type": "string"
          },
          "hours_alone": {
            "maximum": 24.0,
            "minimum": 0.0,
            "title": "Hours Alone",
            "type": "number"
          },
          "housing_type": {
            "enum": [
              "APARTMENT",
              "CONDO",
              "HOUSE"
            ],
            "title": "Housing Type",
            "type": "string"
          },
          "monthly_budget": {
            "maximum": 5000.0,
            "minimum": 0.0,
            "title": "Monthly Budget",
            "type": "number"
          },
          "outdoor_space": {
            "enum": [
              "NONE",
              "BALCONY",
              "YARD"
            ],
            "title": "Outdoor Space",
            "type": "string"
          },
          "rental_allows_pets": {
            "title": "Rental Allows Pets",
            "type": "boolean"
          }
        },
        "required": [
          "housing_type",
          "home_size",
          "rental_allows_pets",
          "outdoor_space",
          "hours_alone",
          "activity_level",
          "grooming_tolerance",
          "monthly_budget",
          "experience",
          "affection_preference"
        ],
        "title": "CatQuestionnaire",
        "type": "object"
      },
      "CatRecommendationRequest": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "const": "CAT",
            "title": "Category",
            "type": "string"
          },
          "destination": {
            "$ref": "#/components/schemas/Destination"
          },
          "questionnaire": {
            "$ref": "#/components/schemas/CatQuestionnaire"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "use_saved_preferences": {
            "default": false,
            "title": "Use Saved Preferences",
            "type": "boolean"
          }
        },
        "required": [
          "destination",
          "session_id",
          "category",
          "questionnaire"
        ],
        "title": "CatRecommendationRequest",
        "type": "object"
      },
      "Category": {
        "enum": [
          "PLANT",
          "DOG",
          "CAT"
        ],
        "title": "Category",
        "type": "string"
      },
...
      "Destination": {
        "additionalProperties": false,
        "properties": {
          "state_code": {
            "anyOf": [
              {
                "pattern": "^[A-Z]{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Code"
          },
          "zip_code": {
            "examples": [
              "10001"
            ],
            "pattern": "^\\d{5}$",
            "title": "Zip Code",
            "type": "string"
          }
        },
        "required": [
          "zip_code"
        ],
        "title": "Destination",
        "type": "object"
      },
      "DogRecommendationRequest": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "const": "DOG",
            "title": "Category",
            "type": "string"
          },
          "destination": {
            "$ref": "#/components/schemas/Destination"
          },
          "questionnaire": {
            "$ref": "#/components/schemas/PetQuestionnaire"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "use_saved_preferences": {
            "default": false,
            "title": "Use Saved Preferences",
            "type": "boolean"
          }
        },
        "required": [
          "destination",
          "session_id",
          "category",
          "questionnaire"
        ],
        "title": "DogRecommendationRequest",
        "type": "object"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "contract_version": {
            "const": "v1",
            "default": "v1",
            "title": "Contract Version",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "LiveAdvisory": {
        "additionalProperties": false,
        "description": "Provider-neutral live context that cannot affect eligibility or score.",
        "properties": {
          "kind": {
            "enum": [
              "CLIMATE",
              "REGULATION",
              "WEB_GUIDANCE"
            ],
            "title": "Kind",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/AdvisorySource"
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Sources",
            "type": "array"
          },
          "summary": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "title",
          "summary",
          "sources"
        ],
        "title": "LiveAdvisory",
        "type": "object"
      },
...
      "PetQuestionnaire": {
        "additionalProperties": false,
        "properties": {
          "activity_level": {
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Activity Level",
            "type": "string"
          },
          "children_present": {
            "default": false,
            "title": "Children Present",
            "type": "boolean"
          },
          "existing_pets": {
            "items": {
              "enum": [
                "DOG",
                "CAT"
              ],
              "type": "string"
            },
            "title": "Existing Pets",
            "type": "array",
            "uniqueItems": true
          },
          "experience": {
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "EXPERT"
            ],
            "title": "Experience",
            "type": "string"
          },
          "grooming_tolerance": {
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Grooming Tolerance",
            "type": "string"
          },
          "home_size": {
            "enum": [
              "SMALL",
              "MEDIUM",
              "LARGE"
            ],
            "title": "Home Size",
            "type": "string"
          },
          "hours_alone": {
            "maximum": 24.0,
            "minimum": 0.0,
            "title": "Hours Alone",
            "type": "number"
          },
          "housing_type": {
            "enum": [
              "APARTMENT",
              "CONDO",
              "HOUSE"
            ],
            "title": "Housing Type",
            "type": "string"
          },
          "monthly_budget": {
            "maximum": 5000.0,
            "minimum": 0.0,
            "title": "Monthly Budget",
            "type": "number"
          },
          "outdoor_space": {
            "enum": [
              "NONE",
              "BALCONY",
              "YARD"
            ],
            "title": "Outdoor Space",
            "type": "string"
          },
          "rental_allows_pets": {
            "title": "Rental Allows Pets",
            "type": "boolean"
          }
        },
        "required": [
          "housing_type",
          "home_size",
          "rental_allows_pets",
          "outdoor_space",
          "hours_alone",
          "activity_level",
          "grooming_tolerance",
          "monthly_budget",
          "experience"
        ],
        "title": "PetQuestionnaire",
        "type": "object"
      },
      "PlantQuestionnaire": {
        "additionalProperties": false,
        "properties": {
          "available_space": {
            "enum": [
              "SMALL",
              "MEDIUM",
              "LARGE"
            ],
            "title": "Available Space",
            "type": "string"
          },
          "children_present": {
            "default": false,
            "title": "Children Present",
            "type": "boolean"
          },
          "experience": {
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "EXPERT"
            ],
            "title": "Experience",
            "type": "string"
          },
          "humidity": {
            "enum": [
              "LOW",
              "AVERAGE",
              "HIGH"
            ],
            "title": "Humidity",
            "type": "string"
          },
          "indoor_temperature_f": {
            "maximum": 95.0,
            "minimum": 50.0,
            "title": "Indoor Temperature F",
            "type": "integer"
          },
          "light_level": {
            "enum": [
              "LOW",
              "MEDIUM",
              "BRIGHT_INDIRECT",
              "DIRECT"
            ],
            "title": "Light Level",
            "type": "string"
          },
          "monthly_budget": {
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Monthly Budget",
            "type": "number"
          },
          "pets_present": {
            "items": {
              "enum": [
                "DOG",
                "CAT"
              ],
              "type": "string"
            },
            "title": "Pets Present",
            "type": "array",
            "uniqueItems": true
          },
          "watering_availability": {
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Watering Availability",
            "type": "string"
          }
        },
        "required": [
          "light_level",
          "humidity",
          "indoor_temperature_f",
          "available_space",
          "experience",
          "watering_availability",
          "monthly_budget"
        ],
        "title": "PlantQuestionnaire",
        "type": "object"
      },
      "PlantRecommendationRequest": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "const": "PLANT",
            "title": "Category",
            "type": "string"
          },
          "destination": {
            "$ref": "#/components/schemas/Destination"
          },
          "questionnaire": {
            "$ref": "#/components/schemas/PlantQuestionnaire"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "use_saved_preferences": {
            "default": false,
            "title": "Use Saved Preferences",
            "type": "boolean"
          }
        },
        "required": [
          "destination",
          "session_id",
          "category",
          "questionnaire"
        ],
        "title": "PlantRecommendationRequest",
        "type": "object"
      },
...
      "RecommendationCompletedEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RecommendationResponse"
          },
          "emitted_at": {
            "format": "date-time",
            "title": "Emitted At",
            "type": "string"
          },
          "event": {
            "const": "completed",
            "default": "completed",
            "title": "Event",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          },
          "sequence": {
            "minimum": 1,
            "title": "Sequence",
            "type": "integer"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "sequence",
          "request_id",
          "session_id",
          "emitted_at",
          "data"
        ],
        "title": "RecommendationCompletedEvent",
        "type": "object"
      },
      "RecommendationFailedEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "emitted_at": {
            "format": "date-time",
            "title": "Emitted At",
            "type": "string"
          },
          "event": {
            "const": "failed",
            "default": "failed",
            "title": "Event",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          },
          "sequence": {
            "minimum": 1,
            "title": "Sequence",
            "type": "integer"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "sequence",
          "request_id",
          "session_id",
          "emitted_at",
          "data"
        ],
        "title": "RecommendationFailedEvent",
        "type": "object"
      },
      "RecommendationItem": {
        "additionalProperties": false,
        "properties": {
          "best_match": {
            "title": "Best Match",
            "type": "boolean"
          },
          "care_summary": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Care Summary",
            "type": "array"
          },
          "concerns": {
            "items": {
              "type": "string"
            },
            "title": "Concerns",
            "type": "array"
          },
          "cost": {
            "$ref": "#/components/schemas/CostEstimate"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "minItems": 1,
            "title": "Evidence",
            "type": "array"
          },
          "local_sources": {
            "items": {
              "$ref": "#/components/schemas/LocalSource"
            },
            "maxItems": 3,
            "title": "Local Sources",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "profile": {
            "title": "Profile",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "minItems": 3,
            "title": "Reasons",
            "type": "array"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "string"
          },
          "safety": {
            "$ref": "#/components/schemas/SafetyDisclosure"
          },
          "scientific_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Scientific Name"
          },
          "score": {
            "maximum": 100,
            "minimum": 0,
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "recommendation_id",
          "name",
          "profile",
          "score",
          "best_match",
          "reasons",
          "evidence",
          "care_summary",
          "cost",
          "safety"
        ],
        "title": "RecommendationItem",
        "type": "object"
      },
      "RecommendationProgressEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProgressData"
          },
          "emitted_at": {
            "format": "date-time",
            "title": "Emitted At",
            "type": "string"
          },
          "event": {
            "const": "progress",
            "default": "progress",
            "title": "Event",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          },
          "sequence": {
            "minimum": 1,
            "title": "Sequence",
            "type": "integer"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "sequence",
          "request_id",
          "session_id",
          "emitted_at",
          "data"
        ],
        "title": "RecommendationProgressEvent",
        "type": "object"
      },
      "RecommendationResponse": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "live_advisories": {
            "items": {
              "$ref": "#/components/schemas/LiveAdvisory"
            },
            "maxItems": 12,
            "title": "Live Advisories",
            "type": "array"
          },
          "metadata": {
            "$ref": "#/components/schemas/RequestMetadata"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationItem"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "Recommendations",
            "type": "array"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "validation_status": {
            "$ref": "#/components/schemas/ValidationStatus"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "metadata",
          "session_id",
          "category",
          "recommendations",
          "validation_status"
        ],
        "title": "RecommendationResponse",
        "type": "object"
      },
      "RecommendationStreamEvent": {
        "discriminator": {
          "mapping": {
            "completed": "#/components/schemas/RecommendationCompletedEvent",
            "failed": "#/components/schemas/RecommendationFailedEvent",
            "progress": "#/components/schemas/RecommendationProgressEvent"
          },
          "propertyName": "event"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/RecommendationProgressEvent"
          },
          {
            "$ref": "#/components/schemas/RecommendationCompletedEvent"
          },
          {
            "$ref": "#/components/schemas/RecommendationFailedEvent"
          }
        ]
      },
      "RequestMetadata": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "versions": {
            "$ref": "#/components/schemas/VersionInfo"
          }
        },
        "required": [
          "request_id",
          "generated_at",
          "versions"
        ],
        "title": "RequestMetadata",
        "type": "object"
      },
...
      "ValidationStatus": {
        "enum": [
          "PASSED",
          "DEGRADED"
        ],
        "title": "ValidationStatus",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "GoogleIapAssertion": {
        "description": "Signed assertion injected by Google Cloud Identity-Aware Proxy.",
        "in": "header",
        "name": "X-Goog-IAP-JWT-Assertion",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Versioned application API for deterministic Plant, Dog, and Cat recommendations and explicitly confirmed care plans.",
    "title": "Indoor Plant and Pet Advisor API",
    "version": "1.0.0",
    "x-api-version": "v1",
    "x-error-contract-version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/recommendations": {
      "post": {
        "operationId": "createRecommendation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "CAT": "#/components/schemas/CatRecommendationRequest",
                    "DOG": "#/components/schemas/DogRecommendationRequest",
                    "PLANT": "#/components/schemas/PlantRecommendationRequest"
                  },
                  "propertyName": "category"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/PlantRecommendationRequest"
                  },
                  {
                    "$ref": "#/components/schemas/DogRecommendationRequest"
                  },
                  {
                    "$ref": "#/components/schemas/CatRecommendationRequest"
                  }
                ],
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or eligibility failure"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Required dependency unavailable"
          }
        },
        "security": [
          {
            "GoogleIapAssertion": []
          }
        ],
        "summary": "Create up to three recommendations",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/v1/recommendations/stream": {
      "post": {
        "operationId": "streamRecommendation",
        "parameters": [
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "CAT": "#/components/schemas/CatRecommendationRequest",
                    "DOG": "#/components/schemas/DogRecommendationRequest",
                    "PLANT": "#/components/schemas/PlantRecommendationRequest"
                  },
                  "propertyName": "category"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/PlantRecommendationRequest"
                  },
                  {
                    "$ref": "#/components/schemas/DogRecommendationRequest"
                  },
                  {
                    "$ref": "#/components/schemas/CatRecommendationRequest"
                  }
                ],
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationStreamEvent"
                },
                "x-sse": {
                  "eventField": "event",
                  "idField": "event_id",
                  "resumeHeader": "Last-Event-ID",
                  "retryMilliseconds": 3000
                }
              }
            },
            "description": "Versioned recommendation progress events"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or eligibility failure"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Required dependency unavailable"
          }
        },
        "security": [
          {
            "GoogleIapAssertion": []
          }
        ],
        "summary": "Stream recommendation workflow progress",
        "tags": [
          "recommendations"
        ]
      }
    }
  },
...
}